+++ src/main/java/com/prgrms/ijuju/domain/stock/begin/scheduler/BeginStockScheduler.java
@@ -2,6 +2,7 @@ package com.prgrms.ijuju.domain.stock.begin.scheduler;
 
 import com.prgrms.ijuju.domain.stock.begin.service.BeginStockGptService;
 import com.prgrms.ijuju.domain.stock.begin.service.BeginStockPriceService;
+import com.prgrms.ijuju.domain.stock.begin.service.BeginStockService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -13,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class BeginStockScheduler {
     private final BeginStockPriceService beginStockPriceService;
     private final BeginStockGptService beginStockGptService;
+    private final BeginStockService beginStockService;
 
     @Scheduled(cron = "0 0 0 * * ?")
     public void createWeeklyStockData() {
@@ -21,5 +23,7 @@ public class BeginStockScheduler {
         for (int i = 0; i < 5; i++) {
             beginStockGptService.generateBeginQuiz();
         }
+
+        beginStockService.refreshStockDataCache();
     }
 }
